Adding Move Player Function and some refactoring
