+++ src/utilties/models/Map.java
@@ -1,6 +1,7 @@
 package utilties.models;
 
-import utilties.models.entities.Empty;
+import utilties.entities.Empty;
+import utilties.entities.IEntity;
 
 public class Map {
 
@@ -23,14 +24,26 @@ public class Map {
         return numberOfRows;
     }
 
+    public void setLocation(IEntity iEntity, Location location){
+        Location previousLocation = iEntity.getLocation();
+        Tile previousTile = map[previousLocation.getX()][previousLocation.getY()];
+        iEntity.setLocation(location);
+        map[location.getX()][location.getY()] = previousTile;
+        map[previousLocation.getX()][previousLocation.getY()] = createBlankTile(location);
+    }
+
     private void initializeMap() {
         map = new Tile[numberOfColumns][numberOfRows];
 
         for (int x = 0; x < numberOfColumns; x++) {
             for (int y = 0; y < numberOfRows; y++) {
                 Location location = new Location(x, y);
-                map[x][y] = new Tile(location, new Empty(location));
+                map[x][y] = createBlankTile(location);
             }
         }
+    }
+
+    private Tile createBlankTile(Location location){
+        return new Tile(location, new Empty(location));
     }
 }
